requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.2
django-cleanup==3.2.0
django-cors-headers==2.5.2
django-seed==0.1.9
django-uuidfield==0.5.0
djangorestframework==3.9.2
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.24.1
Faker==1.0.4
fcm-django==0.2.21
greenlet==0.4.15
idna==2.8
jdcal==1.4
kombu==4.5.0
monotonic==1.5
odfpy==1.4.0
openpyxl==2.6.2
pdfkit==0.6.1
Pillow==6.0.0
# pkg-resources==0.0.0
psycopg2-binary==2.7.7
pyfcm==1.4.5
python-dateutil==2.8.0
python-engineio==3.5.0
python-magic==0.4.15
python-socketio==4.0.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0

# django
# djangorestframework
# django-cors-headers
# psycopg2-binary
# Pillow
# celery
# pdfkit					# Html to Pdf, needs sudo apt-get install wkhtmltopdf
# django-cleanup			
# python-socketio			# socketio
# eventlet				# socketio
# python-magic
# fcm-django
# django-seed
# tablib
